+++ src/compas_timber/connections/generic_joint.py
@@ -86,4 +86,3 @@ class GenericPlateJoint(PlateJoint, GenericJoint):
 
     def __init__(self, plate_a=None, plate_b=None, **kwargs):
         super(GenericPlateJoint, self).__init__(plate_a=plate_a, plate_b=plate_b, **kwargs)
-        # TODO: should we make topology and segment_index required to ensure that this is called as result of PlateConnectionSolver.find_topology?