+++ src/compas_timber/connections/solver.py
@@ -4,21 +4,19 @@ import math
 from compas.geometry import Line
 from compas.geometry import Point
 from compas.geometry import Vector
-from compas.geometry import add_vectors
 from compas.geometry import angle_vectors
-from compas.geometry import closest_point_on_line
-from compas.geometry import cross_vectors
 from compas.geometry import distance_point_line
 from compas.geometry import distance_point_point
 from compas.geometry import dot_vectors
 from compas.geometry import intersection_plane_plane
 from compas.geometry import intersection_segment_polyline
 from compas.geometry import is_parallel_line_line
-from compas.geometry import scale_vector
 from compas.geometry import subtract_vectors
 from compas.plugins import pluggable
 from compas.tolerance import TOL
 
+from compas_timber.utils import distance_segment_segment_points
+from compas_timber.utils import get_segment_overlap
 from compas_timber.utils import is_point_in_polyline
 from compas_timber.utils import do_segments_overlap
 from compas_timber.utils import distance_segment_segment
@@ -129,14 +127,12 @@ class ConnectionSolver(object):
         """
         return find_neighboring_elements(beams, inflate_by=max_distance) if rtree else itertools.combinations(beams, 2)
 
-    def find_topology(self, beam_a, beam_b, tol=TOLERANCE, max_distance=None):
+    def find_topology(self, beam_a, beam_b, max_distance=None):
         """If `beam_a` and `beam_b` intersect within the given `max_distance`, return the topology type of the intersection.
 
         If the topology is role-sensitive, the method outputs the beams in a consistent specific order
         (e.g. main beam first, cross beam second), otherwise, the beams are outputted in the same
         order as they were inputted.
-
-        # TODO: this needs to be reworked ASAP
 
         Parameters
         ----------
@@ -154,101 +150,45 @@ class ConnectionSolver(object):
         tuple(:class:`~compas_timber.connections.JointTopology`, :class:`~compas_timber.parts.Beam`, :class:`~compas_timber.parts.Beam`, float)
 
         """
-        tol = self.TOLERANCE  # TODO: change to a unit-sensitive value
-        angtol = 1e-3
-        a1, a2 = beam_a.centerline
-        b1, b2 = beam_b.centerline
-        va = subtract_vectors(a2, a1)
-        vb = subtract_vectors(b2, b1)
-
-        # check if centerlines parallel
-        ang = angle_vectors(va, vb)
-        if ang < angtol or ang > math.pi - angtol:
-            parallel = True
-        else:
-            parallel = False
-
-        if parallel:
-            pa = a1
-            pb = closest_point_on_line(a1, [b1, b2])
-            exceed, _ = self._exceed_max_distance(pa, pb, max_distance, tol)
-            if exceed:
-                return JointTopology.TOPO_UNKNOWN, None, None, None, None
-
-            # check if any ends meet
-            comb = [[0, 0], [0, 1], [1, 0], [1, 1]]
-            meet = []
-            distance = 0.0
-            
-            for ia, ib in comb:
-                exceed, d = self._exceed_max_distance([a1, a2][ia], [b1, b2][ib], max_distance=max_distance, tol=tol)
-                meet.append(not exceed)
-                if not exceed:
-                    distance = max(distance, d)
-            if sum(meet) != 1:
-                return JointTopology.TOPO_UNKNOWN, None, None, None, None
-
-            # check if overlap: find meeting ends -> compare vectors outgoing from these points
-            meeting_ends_idx = [c for c, m in zip(comb, meet) if m is True][0]
-            ia, ib = meeting_ends_idx
-            pa1 = [a1, a2][ia]
-            pa2 = [a1, a2][not ia]
-            pb1 = [b1, b2][ib]
-            pb2 = [b1, b2][not ib]
-            vA = subtract_vectors(pa2, pa1)
-            vB = subtract_vectors(pb2, pb1)
-            ang = angle_vectors(vA, vB)
-
-            if ang < tol and do_segments_overlap(Line(pa1, pa2), Line(pb1, pb2)):
-                # vectors pointing in the same direction -> beams are overlapping
-                return JointTopology.TOPO_UNKNOWN, None, None, None, None
-            else:
-                return JointTopology.TOPO_I, beam_a, beam_b, distance, (pa1+pb1)/2
-
-        # if not parallel:
-        vn = cross_vectors(va, vb)
-        vna = cross_vectors(va, vn)
-        vnb = cross_vectors(vb, vn)
-
-        ta = self._calc_t([a1, a2], [b1, vnb])
-        pa = Point(*add_vectors(a1, scale_vector(va, ta)))
-        tb = self._calc_t([b1, b2], [a1, vna])
-        pb = Point(*add_vectors(b1, scale_vector(vb, tb)))
-
-        # for max_distance calculations, limit intersection point to line segment
-        if ta < 0:
-            pa = a1
-        if ta > 1:
-            pa = a2
-        if tb < 0:
-            pb = b1
-        if tb > 1:
-            pb = b2
-
-        exceed, distance = self._exceed_max_distance(pa, pb, max_distance, tol)
-        if exceed:
-            print("distance exceeded somehow")
+        # first check if the beams are close enough to be considered intersecting and get the closest points on the segments
+        max_distance = max_distance or TOL.absolute  # TODO: change to a unit-sensitive value
+        dist, point_a, point_b = distance_segment_segment_points(beam_a.centerline, beam_b.centerline)
+        if dist > max_distance:
             return JointTopology.TOPO_UNKNOWN, None, None, None, None
+        point_a = Point(*point_a)
+        point_b = Point(*point_b)
 
-        # topologies:
-        xa = self._is_near_end(ta, beam_a.centerline.length, max_distance or 0, tol)
-        xb = self._is_near_end(tb, beam_b.centerline.length, max_distance or 0, tol)
+        # see if beams are parallel
+        if TOL.is_zero(angle_vectors(beam_a.centerline.direction, beam_b.centerline.direction) % math.pi):
+            # beams are parallel
+            # if parallel overlap on beam_a means that beam_b is overlapped by beam_a. Only need to perform the check on beam_a
+            overlap_on_a = get_segment_overlap(beam_a.centerline, beam_b.centerline)
+            if overlap_on_a is None:
+                return JointTopology.TOPO_I, beam_a, beam_b, dist, (point_a + point_b) / 2.0
+            if overlap_on_a[1] < max_distance:  # overlaps on beam_a start
+                pt = beam_b.endpoint_closest_to_point(beam_a.centerline.start)[1]
+                dist = distance_point_point(pt, beam_a.centerline.start)
+                return JointTopology.TOPO_I, beam_a, beam_b, dist, (beam_a.centerline.start + pt) / 2.0
+            if abs(overlap_on_a[0] - beam_a.length) < max_distance:  # overlaps on beam_a end
+                pt = beam_b.endpoint_closest_to_point(beam_a.centerline.end)[1]
+                dist = distance_point_point(pt, beam_a.centerline.end)
+                return JointTopology.TOPO_I, beam_a, beam_b, dist, (beam_a.centerline.start + pt) / 2.0
+            else:
+                return JointTopology.TOPO_UNKNOWN, None, None, None, None
 
-        pt = (pa+pb)/2.0
-        # L-joint (both meeting at ends)
-        if xa and xb:
-            return JointTopology.TOPO_L, beam_a, beam_b, distance, pt
+        _, a_end_pt = beam_a.endpoint_closest_to_point(point_b)
+        _, b_end_pt = beam_b.endpoint_closest_to_point(point_a)
 
-        # T-joint (one meeting with the end along the other)
-        if xa:
-            # A:main, B:cross
-            return JointTopology.TOPO_T, beam_a, beam_b, distance, pt
-        if xb:
-            # B:main, A:cross
-            return JointTopology.TOPO_T, beam_b, beam_a, distance, pt
-
-        # X-joint (both meeting somewhere along the line)
-        return JointTopology.TOPO_X, beam_a, beam_b, distance, pt
+        a_end = distance_point_point(a_end_pt, point_a) < max_distance
+        b_end = distance_point_point(b_end_pt, point_b) < max_distance
+        location = (point_a + point_b) / 2.0
+        if a_end and b_end:
+            return JointTopology.TOPO_L, beam_a, beam_b, dist, location
+        if a_end:
+            return JointTopology.TOPO_T, beam_a, beam_b, dist, location
+        if b_end:
+            return JointTopology.TOPO_T, beam_b, beam_a, dist, location
+        return JointTopology.TOPO_X, beam_a, beam_b, dist, location
 
     def find_wall_wall_topology(self, wall_a, wall_b, tol=TOLERANCE, max_distance=None):
         """Calculates the topology of the intersection between two walls.