+++ src/compas_timber/design/wall_from_surface.py
@@ -272,7 +272,7 @@ class SurfaceModel(object):
         found_pairs = solver.find_intersecting_pairs(list(model.beams), rtree=True, max_distance=self.dist_tolerance)
         for pair in found_pairs:
             beam_a, beam_b = pair
-            detected_topo, beam_a, beam_b, _ = solver.find_topology(beam_a, beam_b, max_distance=self.dist_tolerance)
+            detected_topo, beam_a, beam_b, _, _ = solver.find_topology(beam_a, beam_b, max_distance=self.dist_tolerance)
             if not detected_topo == JointTopology.TOPO_UNKNOWN:
                 topologies.append({"detected_topo": detected_topo, "beam_a": beam_a, "beam_b": beam_b})
                 for rule in self.rules: