+++ tests/compas_timber/test_plate_joint.py
@@ -130,6 +130,22 @@ def test_simple_joint_and_reset_no_kwargs():
     polyline_b = Polyline([Point(0, 10, 0), Point(10, 10, 0), Point(20, 20, 10), Point(0, 20, 10), Point(0, 10, 0)])
     plate_b = Plate.from_outline_thickness(Polyline(polyline_b.points), 1)
 
+    kwargs = {"topology": JointTopology.TOPO_L, "a_segment_index": 1, "b_segment_index": 0}
+    joint = PlateMiterJoint(plate_a, plate_b, **kwargs)
+    joint.add_features()
+    assert isinstance(joint, PlateMiterJoint), "Expected joint to be a PlateMiterJoint"
+    assert any([plate_a.outline_a.points[i] != polyline_a.points[i] for i in range(len(plate_a.outline_a.points))]), "Expected joint to change outline_a"
+    plate_a.reset()
+    assert all([plate_a.outline_a.points[i] == polyline_a.points[i] for i in range(len(plate_a.outline_a.points))]), "Expected joint to reset outline_a"
+
+
+def test_simple_joint_and_reset_no_kwargs():
+    polyline_a = Polyline([Point(0, 0, 0), Point(0, 10, 0), Point(10, 10, 0), Point(10, 0, 0), Point(0, 0, 0)])
+    plate_a = Plate.from_outline_thickness(Polyline([pt for pt in polyline_a.points]), 1)
+
+    polyline_b = Polyline([Point(0, 10, 0), Point(10, 10, 0), Point(20, 20, 10), Point(0, 20, 10), Point(0, 10, 0)])
+    plate_b = Plate.from_outline_thickness(Polyline(polyline_b.points), 1)
+
     joint = PlateMiterJoint(plate_a, plate_b)
     joint.add_features()
     assert isinstance(joint, PlateMiterJoint), "Expected joint to be a PlateMiterJoint"
@@ -251,7 +267,7 @@ def test_three_plate_joints():
         elif tr[0] == JointTopology.TOPO_EDGE_EDGE:
             joints.append(PlateMiterJoint(tr[1][0], tr[2][0], **kwargs))
         elif tr[0] == JointTopology.TOPO_EDGE_FACE:
-            joints.append(PlateTButtJoint(tr[1][0], tr[2][0], **kwargs))
+            joints.append(PlateButtJoint(tr[1][0], tr[2][0], **kwargs))
 
     assert len(joints) == 3, "Expected three joints"
     assert all(isinstance(j, PlateMiterJoint) for j in joints), "Expected L-joints to be PlateMiterJoint"
@@ -276,12 +292,13 @@ def test_three_plate_joints_mix_topo():
     joints = []
     for tr in topo_results:
         kwargs = {"topology": tr[0], "a_segment_index": tr[1][1], "b_segment_index": tr[2][1]}
+
         if tr[0] == JointTopology.TOPO_UNKNOWN:
             continue
         elif tr[0] == JointTopology.TOPO_EDGE_EDGE:
             joints.append(PlateMiterJoint(tr[1][0], tr[2][0], **kwargs))
         elif tr[0] == JointTopology.TOPO_EDGE_FACE:
-            joints.append(PlateTButtJoint(tr[1][0], tr[2][0], **kwargs))
+            joints.append(PlateButtJoint(tr[1][0], tr[2][0], **kwargs))
 
     assert len(joints) == 3, "Expected three joints"
     assert isinstance(joints[0], PlateTButtJoint), "Expected L-joints to be PlateButtJoint"